Create discord-button.py
This code provides a simple Discord bot that responds to the command "!question" by sending a message with an embedded question and two buttons for the user to choose from. Depending on which button the user clicks, the bot sends a corresponding response.

You can use this code as a starting point for creating interactive features in your Discord server, such as quizzes, polls, or interactive commands.